requirements.txt
discord.py==2.3.2
requests==2.31.0
audioop-lts==0.2.1
aiohttp==3.9.5
aiosqlite==0.20.0
